moved Runner class to separate submodule
